Add product to order connection

+++ Models/Product.cs
@@ -8,6 +8,6 @@ public class Product
     public int ProductId { get; set; }
     public string ProductName { get; set; }
     public string ProductDescription { get; set; }
-    public float ProductPrice { get; set; }
+    public double ProductPrice { get; set; }
     public string ProductStatus { get; set; }
 }
